finalized branch features

+++ src/main/java/io/github/lordship/meterbills/internal/MeterBillsCreateRequest.java
@@ -3,6 +3,7 @@ package io.github.lordship.meterbills.internal;
 import io.github.lordship.meters.MeterMeasurement;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -11,10 +12,10 @@ public record MeterBillsCreateRequest(
         UUID billedMeter,
 
         @NotNull
-        Integer billedAmount,
+        BigDecimal billedAmount,
 
         @NotNull
-        Double rateAmount,
+        BigDecimal rateAmount,
 
         @NotNull
         MeterMeasurement rateUnit,
