+++ src/main/java/io/github/lordship/meterbills/internal/MeterBillsResponse.java
@@ -3,14 +3,15 @@ package io.github.lordship.meterbills.internal;
 import io.github.lordship.meterbills.MeterBills;
 import io.github.lordship.meters.MeterMeasurement;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
 public record MeterBillsResponse(
         UUID uuid,
         UUID billedMeter,
-        int billedAmount,
-        double rateAmount,
+        BigDecimal billedAmount,
+        BigDecimal rateAmount,
         MeterMeasurement rateUnit,
         LocalDate periodStart,
         LocalDate periodEnd
